enhance home screen fix Fetching  jobs from api increase connection timeout to 10sec

+++ app/src/main/java/com/samuelokello/kazihub/presentation/worker/data/Job.kt
@@ -8,7 +8,7 @@ data class Job(
     val id: Int?,
     val title: String?,
     val desc: String?,
-    val budget: Double?,
+    val budget: Int?,
     val location: String?,
     val postedOn: String?,
     val status: Status?,
